requirements.txt
streamlit
langchain>=0.3,<1.0
langchain-openai
langchain-google-genai
langchain-community>=0.3,<1.0
python-dotenv
faiss-cpu
beautifulsoup4
requests